Method NumRound (NEED RESOLVING)

+++ src/Rounding.java
@@ -9,7 +9,7 @@ public class Rounding {
     public double NumRound(double num, int place) {
         number = num;
         decimal = place;
-        number = Math.round(number * Math.pow(10, decimal)) / Math.pow(10, decimal);
+        number = Math.round(number * (double) Math.pow(10, decimal)) / (double) Math.pow(10, decimal);
         return decimal;
     }
 }
